yaml_generation_service/requirements.txt
PyYAML==5.4.1
pika==1.2.0
